+++ extensions/obsidian/src/components/NoteListObsidian.tsx
@@ -3,38 +3,28 @@ import { useEffect, useMemo, useState } from "react";
 import fs from "fs";
 
 import NoteLoader from "../utils/NoteLoader";
-import { Note, SearchNotePreferences } from "../utils/interfaces";
+import { Note, Vault, SearchNotePreferences } from "../utils/interfaces";
 import { NoteList } from "./NoteList";
-import { unpinNote } from "../utils/PinNoteUtils";
 import { filterNotes } from "../utils/utils";
 import { MAX_RENDERED_NOTES } from "../utils/constants";
 
-export function NoteListObsidian(props: { vaultPath: string }) {
-  function unpinNoteAction(note: Note) {
-    return (
-      <Action
-        title="Unpin Note"
-        shortcut={{ modifiers: ["opt", "cmd"], key: "u" }}
-        onAction={() => {
-          const pinnedNotes = unpinNote(note, props.vaultPath);
-        }}
-        icon={Icon.XmarkCircle}
-      />
-    );
-  }
-
+export function NoteListObsidian(props: { vault: Vault }) {
   const pref: SearchNotePreferences = getPreferenceValues();
 
-  const vaultPath = props.vaultPath;
+  const vault = props.vault;
   const [notes, setNotes] = useState<Note[]>([]);
   const [input, setInput] = useState<string>("");
   const list = useMemo(() => filterNotes(notes, input, pref.searchContent), [notes, input]);
 
+  function onDelete(note: Note) {
+    setNotes(notes.filter((n) => n.path !== note.path));
+  }
+
   useEffect(() => {
     async function fetch() {
       try {
-        await fs.promises.access(vaultPath + "/.");
-        const nl = new NoteLoader(vaultPath);
+        await fs.promises.access(vault.path + "/.");
+        const nl = new NoteLoader(vault);
         const _notes = nl.loadNotes();
 
         setNotes(_notes);
@@ -50,11 +40,6 @@ export function NoteListObsidian(props: { vaultPath: string }) {
   }, []);
 
   return (
-    <NoteList
-      notes={list.slice(0, MAX_RENDERED_NOTES)}
-      vaultPath={props.vaultPath}
-      action={unpinNoteAction}
-      onSearchChange={setInput}
-    />
+    <NoteList notes={list.slice(0, MAX_RENDERED_NOTES)} vault={vault} onSearchChange={setInput} onDelete={onDelete} />
   );
 }