+++ extensions/obsidian/src/createNoteCommand.tsx
@@ -1,10 +1,11 @@
-import { List, showToast, Toast } from "@raycast/api";
+import { List } from "@raycast/api";
 
 import { CreateNoteForm } from "./components/CreateNoteForm";
 import { VaultSelection } from "./components/VaultSelection";
 import { Vault } from "./utils/interfaces";
 import { useObsidianVaults } from "./utils/utils";
 import { NoVaultFoundMessage } from "./components/NoVaultFoundMessage";
+import { pathErrorToast } from "./components/Toasts";
 
 export default function Command() {
   const { vaults, ready } = useObsidianVaults();
@@ -12,16 +13,12 @@ export default function Command() {
   if (!ready) {
     return <List isLoading={true}></List>;
   } else if (vaults.length === 0) {
-    return <NoVaultFoundMessage></NoVaultFoundMessage>;
+    return <NoVaultFoundMessage />;
   } else if (vaults.length > 1) {
-    return <VaultSelection vaults={vaults} target={(vault: Vault) => <CreateNoteForm vaultPath={vault.path} />} />;
+    return <VaultSelection vaults={vaults} target={(vault: Vault) => <CreateNoteForm vault={vault} />} />;
   } else if (vaults.length == 1) {
-    return <CreateNoteForm vaultPath={vaults[0].path} />;
+    return <CreateNoteForm vault={vaults[0]} />;
   } else {
-    showToast({
-      title: "Path Error",
-      message: "Something went wrong with your vault path. There are no paths to select from.",
-      style: Toast.Style.Failure,
-    });
+    pathErrorToast();
   }
 }