+++ extensions/obsidian/src/utils/NoteCreator.tsx
@@ -1,9 +1,9 @@
 import { showToast, Toast, confirmAlert, Icon, open, getPreferenceValues, Clipboard } from "@raycast/api";
-
 import path from "path";
 import fs from "fs";
-import { NoteFormPreferences, FormValue } from "./interfaces";
-import { monthMapping, dayMapping } from "./utils";
+
+import { NoteFormPreferences, FormValue, Vault } from "./interfaces";
+import { monthMapping, dayMapping, applyTemplates } from "./utils";
 
 class NoteCreator {
   vaultPath: string;
@@ -11,19 +11,22 @@ class NoteCreator {
   saved = false;
   pref: NoteFormPreferences;
 
-  constructor(noteProps: FormValue, vaultPath: string, pref: NoteFormPreferences) {
-    this.vaultPath = vaultPath;
+  constructor(noteProps: FormValue, vault: Vault, pref: NoteFormPreferences) {
+    this.vaultPath = vault.path;
     this.noteProps = noteProps;
     this.pref = pref;
   }
 
-  createNote() {
+  async createNote() {
     if (this.noteProps.name == "") {
       this.noteProps.name = this.pref.prefNoteName;
     }
-    const content = this.buildNoteContent();
-    const name = this.applyTemplates(this.noteProps.name);
+    let content = this.addYAMLFrontmatter("");
+    content = await applyTemplates(content);
+    const name = await applyTemplates(this.noteProps.name);
+
     this.saveNote(content, name);
+
     if (this.pref.openOnCreate) {
       const target =
         "obsidian://open?path=" + encodeURIComponent(path.join(this.vaultPath, this.noteProps.path, name + ".md"));
@@ -32,34 +35,7 @@ class NoteCreator {
     return this.saved;
   }
 
-  applyTemplates(content: string) {
-    const date = new Date();
-    const hours = date.getHours().toString().padStart(2, "0");
-    const minutes = date.getMinutes().toString().padStart(2, "0");
-    const seconds = date.getSeconds().toString().padStart(2, "0");
-
-    const timestamp = Date.now().toString();
-
-    content = content.replaceAll("{time}", date.toLocaleTimeString());
-    content = content.replaceAll("{date}", date.toLocaleDateString());
-
-    content = content.replaceAll("{year}", date.getFullYear().toString());
-    content = content.replaceAll("{month}", monthMapping[date.getMonth()]);
-    content = content.replaceAll("{day}", dayMapping[date.getDay()]);
-
-    content = content.replaceAll("{hour}", hours);
-    content = content.replaceAll("{minute}", minutes);
-    content = content.replaceAll("{second}", seconds);
-    content = content.replaceAll("{millisecond}", date.getMilliseconds().toString());
-
-    content = content.replaceAll("{timestamp}", timestamp);
-    content = content.replaceAll("{zettelkastenID}", timestamp);
-
-    return content;
-  }
-
-  buildNoteContent() {
-    let content = "";
+  addYAMLFrontmatter(content: string) {
     if (this.noteProps.tags.length > 0) {
       content = "---\ntags: [";
       for (let i = 0; i < this.noteProps.tags.length - 1; i++) {
@@ -68,7 +44,6 @@ class NoteCreator {
       content += '"' + this.noteProps.tags.pop() + '"]\n---\n';
     }
     content += this.noteProps.content;
-    content = this.applyTemplates(content);
 
     return content;
   }