+++ extensions/obsidian/src/utils/PinNoteUtils.tsx
@@ -1,7 +1,7 @@
-import { showToast, Toast, environment } from "@raycast/api";
+import { showToast, Toast, environment, Icon, confirmAlert } from "@raycast/api";
 import fs from "fs";
 
-import { Note } from "./interfaces";
+import { Note, Vault } from "./interfaces";
 import { getNoteFileContent } from "./utils";
 
 interface DataProps {
@@ -9,7 +9,28 @@ interface DataProps {
   pinnedNotes: Note[];
 }
 
-function getInfo(vaultPath: string) {
+export async function resetPinnedNotes(vault: Vault) {
+  const info = getInfo(vault);
+
+  info.vault.pinnedNotes = [];
+
+  const idx = info.data.findIndex((v) => v.vaultPath == vault.path);
+  info.data[idx] = info.vault;
+
+  const options = {
+    title: "Reset Pinned Notes",
+    message: 'Are you sure you want to reset all pinned notes for: "' + vault.name + '"?',
+    icon: Icon.ExclamationMark,
+  };
+
+  if (await confirmAlert(options)) {
+    fs.writeFileSync(environment.supportPath + "/data.json", JSON.stringify(info.data));
+    showToast(Toast.Style.Success, "Reset Pinned Notes for " + vault.name);
+    return true;
+  }
+}
+
+function getInfo(vault: Vault) {
   if (!fs.existsSync(environment.supportPath + "/data.json")) {
     fs.writeFileSync(environment.supportPath + "/data.json", "[]");
   }
@@ -17,9 +38,9 @@ function getInfo(vaultPath: string) {
   const data = fs.readFileSync(environment.supportPath + "/data.json", "utf8");
   const parsedData: DataProps[] = JSON.parse(data);
 
-  const vaults = parsedData.filter((vault) => vault.vaultPath == vaultPath);
+  const vaults = parsedData.filter((v) => v.vaultPath == vault.path);
   if (vaults.length == 0) {
-    const emptyData = { vaultPath: vaultPath, pinnedNotes: [] };
+    const emptyData = { vaultPath: vault.path, pinnedNotes: [] };
     parsedData.push(emptyData);
     fs.writeFileSync(environment.supportPath + "/data.json", JSON.stringify(parsedData));
     return { vault: emptyData, pinnedNotes: emptyData.pinnedNotes, data: parsedData };
@@ -28,27 +49,27 @@ function getInfo(vaultPath: string) {
   }
 }
 
-export function getPinnedNotes(path: string): Note[] {
-  const info = getInfo(path);
+export function getPinnedNotes(vault: Vault): Note[] {
+  const info = getInfo(vault);
 
   // Make sure old pinned notes conform to newest interface
   info.pinnedNotes.forEach((note) => {
     if (note.content == undefined || note.content == "") {
       note.content = getNoteFileContent(note.path);
-      unpinNote(note, path);
-      pinNote(note, path);
+      unpinNote(note, vault);
+      pinNote(note, vault);
     }
   });
   return info.pinnedNotes;
 }
 
-export function unpinNote(note: Note, vaultPath: string) {
-  const info = getInfo(vaultPath);
+export function unpinNote(note: Note, vault: Vault) {
+  const info = getInfo(vault);
 
-  const pinnedNotes = info.pinnedNotes.filter((pinnedNote) => pinnedNote.key !== note.key);
+  const pinnedNotes = info.pinnedNotes.filter((pinnedNote) => pinnedNote.path !== note.path);
   info.vault.pinnedNotes = pinnedNotes;
 
-  const idx = info.data.findIndex((vault) => vault.vaultPath == vaultPath);
+  const idx = info.data.findIndex((v) => v.vaultPath == vault.path);
   info.data[idx] = info.vault;
 
   fs.writeFileSync(environment.supportPath + "/data.json", JSON.stringify(info.data));
@@ -62,15 +83,15 @@ export function unpinNote(note: Note, vaultPath: string) {
   return pinnedNotes;
 }
 
-export function isNotePinned(note: Note, vaultPath: string) {
-  const info = getInfo(vaultPath);
-  return info.pinnedNotes.filter((pinnedNote) => pinnedNote.key == note.key).length !== 0;
+export function isNotePinned(note: Note, vault: Vault) {
+  const info = getInfo(vault);
+  return info.pinnedNotes.filter((pinnedNote) => pinnedNote.path == note.path).length !== 0;
 }
 
-export function pinNote(note: Note, vaultPath: string) {
-  const info = getInfo(vaultPath);
+export function pinNote(note: Note, vault: Vault) {
+  const info = getInfo(vault);
 
-  if (isNotePinned(note, vaultPath)) {
+  if (isNotePinned(note, vault)) {
     showToast({
       title: "Already Pinned",
       message: "'" + note.title + "' is already pinned.",
@@ -81,7 +102,7 @@ export function pinNote(note: Note, vaultPath: string) {
 
   info.vault.pinnedNotes.push(note);
 
-  const idx = info.data.findIndex((vault) => vault.vaultPath == vaultPath);
+  const idx = info.data.findIndex((v) => v.vaultPath == vault.path);
   info.data[idx] = info.vault;
 
   fs.writeFileSync(environment.supportPath + "/data.json", JSON.stringify(info.data));