+++ extensions/obsidian/src/utils/actions.tsx
@@ -7,19 +7,19 @@ import {
   showToast,
   Toast,
   getSelectedText,
+  Color,
 } from "@raycast/api";
 
 import fs from "fs";
 import React, { useState } from "react";
 
 import { AppendNoteForm } from "../components/AppendNoteForm";
-import { SearchNotePreferences, Note } from "./interfaces";
+import { EditNote } from "../components/EditNote";
+import { SearchNotePreferences, Note, Vault } from "./interfaces";
 import { isNotePinned, pinNote, unpinNote } from "./PinNoteUtils";
-
-enum PrimaryAction {
-  QuickLook = "quicklook",
-  OpenInObsidian = "obsidian",
-}
+import { NoteQuickLook } from "../components/NoteQuickLook";
+import { deleteNote } from "./utils";
+import { NoteAction, PrimaryAction } from "./constants";
 
 async function appendSelectedTextTo(note: Note) {
   const pref: SearchNotePreferences = getPreferenceValues();
@@ -44,24 +44,13 @@ async function appendSelectedTextTo(note: Note) {
   }
 }
 
-function NoteQuickLook(props: { note: Note; vaultPath: string }) {
+export function NoteActions(props: { note: Note; vault: Vault; actionCallback: (action: NoteAction) => void }) {
   const note = props.note;
-  return (
-    <Detail
-      markdown={note.content}
-      actions={
-        <ActionPanel>
-          <Action.Open title="Open in Obsidian" target={"obsidian://open?path=" + encodeURIComponent(note.path)} />
-        </ActionPanel>
-      }
-    />
-  );
-}
+  const vault = props.vault;
+  const URIEncodedPath = encodeURIComponent(note.path);
+  const actionCallback = props.actionCallback;
 
-export function NoteActions(props: { note: Note; vaultPath: string; onPin: () => void }) {
-  const note = props.note;
-
-  const [pinned, setPinned] = useState(isNotePinned(note, props.vaultPath));
+  const [pinned, setPinned] = useState(isNotePinned(note, vault));
 
   return (
     <React.Fragment>
@@ -73,8 +62,15 @@ export function NoteActions(props: { note: Note; vaultPath: string; onPin: () =>
       />
 
       <Action.Push
+        title="Edit Note"
+        target={<EditNote note={note} actionCallback={actionCallback} />}
+        shortcut={{ modifiers: ["opt"], key: "e" }}
+        icon={Icon.Pencil}
+      />
+
+      <Action.Push
         title="Append to Note"
-        target={<AppendNoteForm note={note} />}
+        target={<AppendNoteForm note={note} actionCallback={actionCallback}/>}
         shortcut={{ modifiers: ["opt"], key: "a" }}
         icon={Icon.Pencil}
       />
@@ -99,14 +95,14 @@ export function NoteActions(props: { note: Note; vaultPath: string; onPin: () =>
       <Action.CopyToClipboard
         title="Copy Markdown Link"
         icon={Icon.Link}
-        content={`[${note.title}](obsidian://open?path=${encodeURIComponent(note.path)})`}
+        content={`[${note.title}](obsidian://open?path=${URIEncodedPath})`}
         shortcut={{ modifiers: ["opt"], key: "l" }}
       />
 
       <Action.CopyToClipboard
         title="Copy Obsidian URI"
         icon={Icon.Link}
-        content={`obsidian://open?path=${encodeURIComponent(note.path)}`}
+        content={`obsidian://open?path=${URIEncodedPath}`}
         shortcut={{ modifiers: ["opt"], key: "u" }}
       />
 
@@ -115,30 +111,45 @@ export function NoteActions(props: { note: Note; vaultPath: string; onPin: () =>
         shortcut={{ modifiers: ["opt"], key: "p" }}
         onAction={() => {
           if (pinned) {
-            unpinNote(note, props.vaultPath);
+            unpinNote(note, vault);
             setPinned(!pinned);
-            props.onPin();
+            actionCallback(NoteAction.Pin);
           } else {
-            pinNote(note, props.vaultPath);
+            pinNote(note, vault);
             setPinned(!pinned);
-            props.onPin();
+            actionCallback(NoteAction.Pin);
           }
         }}
         icon={pinned ? Icon.XmarkCircle : Icon.Pin}
+      />
+
+      <Action
+        title="Delete Note"
+        shortcut={{ modifiers: ["opt"], key: "d" }}
+        onAction={() => {
+          const deleted = deleteNote(note, vault);
+          deleted.then((d) => {
+            if (d) {
+              actionCallback(NoteAction.Delete);
+            }
+          });
+        }}
+        icon={{ source: Icon.Trash, tintColor: Color.Red }}
       />
     </React.Fragment>
   );
 }
 
-export function OpenNoteActions(props: { note: Note; vaultPath: string }) {
+export function OpenNoteActions(props: { note: Note; vault: Vault; actionCallback: (action: NoteAction) => void }) {
   const note = props.note;
+  const vault = props.vault;
   const pref: SearchNotePreferences = getPreferenceValues();
   const primaryAction = pref.primaryAction;
 
   const quicklook = (
     <Action.Push
       title="Quick Look"
-      target={<NoteQuickLook note={note} vaultPath={props.vaultPath} />}
+      target={<NoteQuickLook note={note} vault={vault} actionCallback={props.actionCallback} />}
       icon={Icon.Eye}
     />
   );